Jarvis 3.0 Secured

+++ backend/server.js
@@ -3,6 +3,7 @@ import cors from 'cors'
 import dotenv from 'dotenv'
 import axios from 'axios'
 import mongoose from 'mongoose'
+import jwt from 'jsonwebtoken'
 
 dotenv.config();
 const app = express();
@@ -44,6 +45,14 @@ let name = "";
 
 app.post('/conversations', async (req, res) => {
   try {
+    const token = req.headers.authorization?.split(' ')[1];
+    if (!token) {
+      return res.status(401).json({ error: 'Unauthorized' });
+    }
+    const decoded = jwt.verify(token, process.env.JWT_SECRET);
+    if (!decoded || !decoded.username) {
+      return res.status(401).json({ error: 'Invalid token' });
+    }
     const { sender, message, timestamp, conversationId, username } = req.body;
     name = username;
     const newMessage = { sender, message, timestamp, username };
@@ -65,6 +74,14 @@ app.post('/conversations', async (req, res) => {
 app.post('/api/gemini', async (req, res) => {
 
   try {
+    const token = req.headers.authorization?.split(' ')[1];
+    if (!token) {
+      return res.status(401).json({ error: 'Unauthorized' });
+    }
+    const decoded = jwt.verify(token, process.env.JWT_SECRET);
+    if (!decoded || !decoded.username) {
+      return res.status(401).json({ error: 'Invalid token' });
+    }
     const { prompt, username } = req.body;
     const convo = await Model.findOne({ username });
 
@@ -101,6 +118,14 @@ app.post('/api/gemini', async (req, res) => {
 
 app.get('/conversations/:username', async (req, res) => {
   const { username } = req.params;
+  const token = req.headers.authorization?.split(' ')[1];
+    if (!token) {
+      return res.status(401).json({ error: 'Unauthorized' });
+    }
+    const decoded = jwt.verify(token, process.env.JWT_SECRET);
+    if (!decoded || !decoded.username) {
+      return res.status(401).json({ error: 'Invalid token' });
+    }
 
   try {
     const userConvo = await Model.findOne({ username: username });
@@ -119,6 +144,14 @@ app.get('/conversations/:username', async (req, res) => {
 app.post('/convoss/:username', async (req, res) => {
   try {
     const { username } = req.params;
+    const token = req.headers.authorization?.split(' ')[1];
+    if (!token) {
+      return res.status(401).json({ error: 'Unauthorized' });
+    }
+    const decoded = jwt.verify(token, process.env.JWT_SECRET);
+    if (!decoded || !decoded.username) {
+      return res.status(401).json({ error: 'Invalid token' });
+    }
     const deleteConvo = await Model.findOneAndDelete({ username });
 
     if (deleteConvo) {
